[BEAM-9273] code review - to be squashed
