+++ runners/apex/src/main/java/org/apache/beam/runners/apex/translation/ParDoTranslator.java
@@ -31,12 +31,12 @@ import java.util.stream.Collectors;
 import org.apache.beam.runners.apex.ApexRunner;
 import org.apache.beam.runners.apex.translation.operators.ApexParDoOperator;
 import org.apache.beam.runners.core.SplittableParDoViaKeyedWorkItems.ProcessElements;
-import org.apache.beam.runners.core.construction.DoFnFeatures;
 import org.apache.beam.runners.core.construction.ParDoTranslation;
 import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.transforms.DoFn;
 import org.apache.beam.sdk.transforms.DoFnSchemaInformation;
 import org.apache.beam.sdk.transforms.ParDo;
+import org.apache.beam.sdk.transforms.reflect.DoFnSignatures;
 import org.apache.beam.sdk.values.PCollection;
 import org.apache.beam.sdk.values.PCollectionView;
 import org.apache.beam.sdk.values.PValue;
@@ -58,18 +58,18 @@ class ParDoTranslator<InputT, OutputT>
   public void translate(ParDo.MultiOutput<InputT, OutputT> transform, TranslationContext context) {
     DoFn<InputT, OutputT> doFn = transform.getFn();
 
-    if (DoFnFeatures.isSplittable(doFn)) {
+    if (DoFnSignatures.isSplittable(doFn)) {
       throw new UnsupportedOperationException(
           String.format(
               "%s does not support splittable DoFn: %s", ApexRunner.class.getSimpleName(), doFn));
     }
-    if (DoFnFeatures.requiresTimeSortedInput(doFn)) {
+    if (DoFnSignatures.requiresTimeSortedInput(doFn)) {
       throw new UnsupportedOperationException(
           String.format(
               "%s doesn't currently support @RequiresTimeSortedInput",
               ApexRunner.class.getSimpleName()));
     }
-    if (DoFnFeatures.usesTimers(doFn)) {
+    if (DoFnSignatures.usesTimers(doFn)) {
       throw new UnsupportedOperationException(
           String.format(
               "Found %s annotations on %s, but %s cannot yet be used with timers in the %s.",
