+++ runners/gearpump/src/main/java/org/apache/beam/runners/gearpump/translators/ParDoMultiOutputTranslator.java
@@ -24,7 +24,6 @@ import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import org.apache.beam.runners.core.construction.DoFnFeatures;
 import org.apache.beam.runners.core.construction.ParDoTranslation;
 import org.apache.beam.runners.gearpump.GearpumpRunner;
 import org.apache.beam.runners.gearpump.translators.functions.DoFnFunction;
@@ -33,6 +32,7 @@ import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.transforms.DoFn;
 import org.apache.beam.sdk.transforms.DoFnSchemaInformation;
 import org.apache.beam.sdk.transforms.ParDo;
+import org.apache.beam.sdk.transforms.reflect.DoFnSignatures;
 import org.apache.beam.sdk.util.WindowedValue;
 import org.apache.beam.sdk.values.PCollection;
 import org.apache.beam.sdk.values.PCollectionView;
@@ -82,7 +82,7 @@ public class ParDoMultiOutputTranslator<InputT, OutputT>
     Map<String, PCollectionView<?>> sideInputMapping =
         ParDoTranslation.getSideInputMapping(context.getCurrentTransform());
 
-    if (DoFnFeatures.requiresTimeSortedInput(transform.getFn())) {
+    if (DoFnSignatures.requiresTimeSortedInput(transform.getFn())) {
       throw new UnsupportedOperationException(
           String.format(
               "%s doesn't currently support @RequiresTimeSortedInput annotation",
