+++ runners/samza/src/main/java/org/apache/beam/runners/samza/translation/ParDoBoundMultiTranslator.java
@@ -29,7 +29,6 @@ import java.util.ServiceLoader;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 import org.apache.beam.model.pipeline.v1.RunnerApi;
-import org.apache.beam.runners.core.construction.DoFnFeatures;
 import org.apache.beam.runners.core.construction.ParDoTranslation;
 import org.apache.beam.runners.core.construction.graph.PipelineNode;
 import org.apache.beam.runners.core.construction.graph.QueryablePipeline;
@@ -97,13 +96,13 @@ class ParDoBoundMultiTranslator<InT, OutT>
             .collect(
                 Collectors.toMap(e -> e.getKey(), e -> ((PCollection<?>) e.getValue()).getCoder()));
 
-    boolean isStateful = DoFnFeatures.isStateful(transform.getFn());
+    boolean isStateful = DoFnSignatures.isStateful(transform.getFn());
     final Coder<?> keyCoder = isStateful ? ((KvCoder<?, ?>) input.getCoder()).getKeyCoder() : null;
 
-    if (DoFnFeatures.isSplittable(transform.getFn())) {
+    if (DoFnSignatures.isSplittable(transform.getFn())) {
       throw new UnsupportedOperationException("Splittable DoFn is not currently supported");
     }
-    if (DoFnFeatures.requiresTimeSortedInput(transform.getFn())) {
+    if (DoFnSignatures.requiresTimeSortedInput(transform.getFn())) {
       throw new UnsupportedOperationException(
           "@RequiresTimeSortedInput annotation is not currently supported");
     }
